+++ src/main/java/cn/monitor4all/logRecord/service/AysLogService.java
@@ -1,0 +1,12 @@
+package cn.monitor4all.logRecord.service;
+
+import org.aspectj.lang.ProceedingJoinPoint;
+
+/***
+ *  异步写入log
+ */
+public interface AysLogService {
+
+    void aysWriteLog(ProceedingJoinPoint pjp, Object result, Boolean isSuccess, String errorMsg, Long executionTime) throws InterruptedException;
+
+}
