+++ src/test/java/cn/monitor4all/logRecord/test/operationLogNameTest/LogRecordFuncTest.java
@@ -14,6 +14,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestComponent;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.context.annotation.Scope;
+import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
@@ -26,6 +27,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @SpringBootTest
 @ContextConfiguration(classes = {LogRecordAutoConfiguration.class, LogRecordFuncTest.TestLogListener.class, OperationLogService.class})
 @EnableAspectJAutoProxy(proxyTargetClass = true)
+@EnableAsync
 public class LogRecordFuncTest {
 
     @Autowired
